--- a/Jint.Repl/Program.cs
+++ b/Jint.Repl/Program.cs
@@ -14,7 +14,6 @@ namespace Jint.Repl
         {
             var engine = new Engine(cfg => cfg
                 .AllowClr()
-                .CollectStackTrace()
             );
 
             engine