--- a/Jint.Tests.Ecma/EcmaTest.cs
+++ b/Jint.Tests.Ecma/EcmaTest.cs
@@ -199,9 +199,6 @@ namespace Jint.Tests.Ecma
             var pacificTimeZone = TimeZoneInfo.FindSystemTimeZoneById("Pacific Standard Time");
             var engine = new Engine(cfg => cfg
                 .LocalTimeZone(pacificTimeZone)
-#if DEBUG
-                .CollectStackTrace()
-#endif
             );
 
             // loading driver