@lahma 2020-12-20
Converting to draft and trying to see if there's some wins for reported issues (missing stack traces and locations).